requirements.txt
sortedcontainers>=1.5.7
sqlalchemy>=1.0.0
requests>=2.12.0
pyyaml>=3.12
psycopg2>=2.6.2